requirements.txt
numpy
pandas
matplotlib
seaborn
tensorflow
scikit-learn
streamlit
pillow